Moved MockDHTProvider to its own file
