+++ ipv8/test/messaging/anonymization/test_community.py
@@ -1,7 +1,7 @@
 from asyncio import Future
-from collections import defaultdict
 from unittest.mock import Mock
 
+from .mock import MockDHTProvider
 from ...base import TestBase
 from ...mocking.endpoint import MockEndpointListener
 from ...mocking.exit_socket import MockTunnelExitSocket
@@ -12,26 +12,6 @@ from ....messaging.anonymization.tunnel import (CIRCUIT_STATE_EXTENDING, PEER_FL
                                                 PEER_FLAG_EXIT_IPV8, PEER_FLAG_SPEED_TEST)
 from ....messaging.interfaces.udp.endpoint import DomainAddress, UDPEndpoint
 from ....util import succeed
-
-# Map of info_hash -> peer list
-global_dht_services = defaultdict(list)
-
-
-class MockDHTProvider(object):
-
-    def __init__(self, peer):
-        self.peer = peer
-        # DHTDiscoveryCommunity functionality
-        global_dht_services[peer.mid].append(peer)
-
-    async def peer_lookup(self, mid, peer=None):
-        return await self.lookup(mid)
-
-    async def lookup(self, info_hash):
-        return info_hash, global_dht_services.get(info_hash, [])
-
-    async def announce(self, info_hash, intro_point):
-        global_dht_services[info_hash].append(intro_point)
 
 
 class TestTunnelCommunity(TestBase):
